requirements.txt

fastapi==0.110.0
uvicorn==0.27.1
boto3==1.34.34
pydantic==2.6.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
